+++ tests/test_clusterings/test_clusterings.py
@@ -2,9 +2,23 @@ import os
 import pytest
 
 from kmeans import kmeans
+from affinity_propagation import affinity_propagation
+from mean_shift import mean_shift
+from dbscan import dbscan
+from optics import optics
+from spectral_clustering import spectral_clustering
+from gaussian_mixture.gaussian_mixture import gaussian_mixture
+from hierarchical_clustering.agglomerative_clustering import agglomerative_clustering
 
 CLUSTERINGS = {
     "KMEANS": [kmeans],
+    "AFFINITY_PROPAGATION": [affinity_propagation],
+    "MEAN_SHIFT": [mean_shift],
+    "DBSCAN": [dbscan],
+    "OPTICS": [optics],
+    "SPECTRAL_CLUSTERING": [spectral_clustering],
+    "GAUSSIAN_MIXTURE": [gaussian_mixture],
+    "HIERARCHICAL_CLUSTERING":[agglomerative_clustering],
 }
 
 @pytest.fixture(scope="session", autouse=True)